requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-environ==0.12.0
django-tailwind==4.2.0
django-htmx==1.24.0

requests==2.32.5
requests-cache==1.2.1
pandas==2.3.2
numpy==2.3.3
scikit-learn==1.7.2
scipy==1.16.1
sentence-transformers==5.1.0
tf_keras==2.20.1